requirements.txt
discord.py[voice]==2.0.1
asyncpraw==7.5.0
prawcore==2.3.0
yt-dlp==2023.3.4
Pillow==9.3.0
python-dotenv==0.20.0
deta==1.1.0
wavelink==2.1.0
pytest==7.3.1
